+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Details panel responds to mouse scroll in all tabs
 - Details panel sets COLUMNS to allow jj diff tool to fit window
 - Update the details panel when gaining focus
+- Added an animated popup for fetch/push operations
 
 ### Changed